Added download link in getting started
The first step of getting started should be download. Either we can mention all the download steps there, or just put a link there. This just puts the link to download page.
I am not sure if this pull request follows the UI coding standards. Please review and merge or discard.
Thanks.