Change "Fork on Github" spacing to match home page
The hope page has a space between the fork icon and the words "Fork on GitHub". Other pages, such as Download, do not. This can be a bit distracting when going back and forth between the pages. I think it looks better with the space, so I'm proposing this change as a fix. It may be minor, but it eliminates a distraction.